refactor: Verification 패키지 분리
