+++ src/main/java/com/cotato/kampus/domain/verification/dao/VerificationPhotoRepository.java
@@ -1,8 +1,8 @@
-package com.cotato.kampus.domain.user.dao;
+package com.cotato.kampus.domain.verification.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.cotato.kampus.domain.user.domain.VerificationPhoto;
+import com.cotato.kampus.domain.verification.domain.VerificationPhoto;
 
 public interface VerificationPhotoRepository extends JpaRepository<VerificationPhoto, Long> {
 }
