+++ src/main/java/com/cotato/kampus/domain/verification/dao/VerificationRecordRepository.java
@@ -1,10 +1,10 @@
-package com.cotato.kampus.domain.user.dao;
+package com.cotato.kampus.domain.verification.dao;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.cotato.kampus.domain.user.domain.VerificationRecord;
+import com.cotato.kampus.domain.verification.domain.VerificationRecord;
 
 public interface VerificationRecordRepository extends JpaRepository<VerificationRecord, Long> {
